Add functionalities to filter by the owner user and create a one to many relationship between users and listings. and also authentication.

+++ routes/web.php
@@ -25,8 +25,7 @@ Route::get('/', [ListingController::class,'index']);
 
 
 //Show form to create new listings
-Route::get('/listings/create',[ListingController::class,'create']);
-
+Route::get('/listings/create',[ListingController::class,'create'])->middleware('auth');
 
 
 //Store listing Data
@@ -34,23 +33,52 @@ Route::post('/listings',[ListingController::class,'store']);
 
 
 //Update Listings
-Route::put('/listings/{listing}',[ListingController::class,'update']);
+Route::put('/listings/{listing}',[ListingController::class,'update'])->middleware('auth');
 
 //Delete Listings
-Route::delete('/listings/{listing}',[ListingController::class,'destroy']);
+Route::delete('/listings/{listing}',[ListingController::class,'destroy'])->middleware('auth');
 
 //show edit form
-Route::get('/listings/{listing}/edit',[ListingController::class,'edit']);
+Route::get('/listings/{listing}/edit',[ListingController::class,'edit'])->middleware('auth');
 
+
+//Manage Listings 
+Route::get('/listings/manage',[ListingController::class,'manage'])->middleware('auth');
 
 //Single Listings
 Route::get('/listings/{listing}',[ListingController::class,'show']);
 
+
+
 //Show the Register Form
-Route::get('/register',[UserController::class,'create']);
+Route::get('/register',[UserController::class,'create'])->middleware('guest');
 
 //Create new user
 Route::post('users',[UserController::class,'store']);
+
+
+//Logout user
+Route::post('/logout',[UserController::class,'logout']);
+
+//Show log in form
+Route::get('/login',[UserController::class,'login'])->name('login')->middleware('guest');
+
+
+//Log in user
+Route::post('/users/authenticate',[UserController::class,'authenticate']);
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 //function(Listing $listing){
 
@@ -103,4 +131,4 @@ Route::post('users',[UserController::class,'store']);
 //     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 // });
 
-require __DIR__.'/auth.php';
+// require __DIR__.'/auth.php';
